Добавил jsp с едой, + мелкие правки.
