+++ src/main/java/ru/javawebinar/topjava/model/Meal.java
@@ -1,29 +1,14 @@
 package ru.javawebinar.topjava.model;
 
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
 import java.time.LocalDateTime;
 
+@Getter
+@RequiredArgsConstructor
 public class Meal {
     private final LocalDateTime dateTime;
-
     private final String description;
-
     private final int calories;
-
-    public Meal(LocalDateTime dateTime, String description, int calories) {
-        this.dateTime = dateTime;
-        this.description = description;
-        this.calories = calories;
-    }
-
-    public LocalDateTime getDateTime() {
-        return dateTime;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public int getCalories() {
-        return calories;
-    }
 }
