+++ src/main/java/ru/javawebinar/topjava/model/MealWithExceed.java
@@ -1,31 +1,24 @@
 package ru.javawebinar.topjava.model;
 
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import lombok.ToString;
+
 import java.time.LocalDateTime;
 
+@ToString
+@Getter
+@RequiredArgsConstructor
 public class MealWithExceed {
     private final LocalDateTime dateTime;
-
     private final String description;
-
     private final int calories;
-
     private final boolean exceed;
-
-    public MealWithExceed(LocalDateTime dateTime, String description, int calories, boolean exceed) {
-        this.dateTime = dateTime;
-        this.description = description;
-        this.calories = calories;
-        this.exceed = exceed;
-    }
 
     public MealWithExceed(Meal meal, boolean exceed) {
         this.dateTime = meal.getDateTime();
         this.description = meal.getDescription();
         this.calories = meal.getCalories();
         this.exceed = exceed;
-    }
-
-    public LocalDateTime getDateTime() {
-        return dateTime;
     }
 }
