+++ src/main/java/ru/javawebinar/topjava/util/TimeUtil.java
@@ -1,9 +1,17 @@
 package ru.javawebinar.topjava.util;
 
+import java.time.LocalDateTime;
 import java.time.LocalTime;
+import java.time.format.DateTimeFormatter;
 
 public class TimeUtil {
+    private static final DateTimeFormatter FORMATTER = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm");
+
     public static boolean isBetween(LocalTime lt, LocalTime startTime, LocalTime endTime) {
         return lt.compareTo(startTime) >= 0 && lt.compareTo(endTime) <= 0;
+    }
+
+    public static String toString(LocalDateTime dateTime) {
+        return dateTime == null ? "" : dateTime.format(FORMATTER);
     }
 }
